+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.3.0 [03/08/2019]
+### Additions
+- README of the repository is in a Paper element below the languages.
+
+## Changes
+- User is now in the title of the card with links.
+
 ## 1.2.0 [03/05/2019]
 ### Additions
 - User under repository name is now a link to their profile page.